requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.12
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==3.4
iniconfig==1.1.1
model-bakery==1.7.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
